Moving Services to Structure Table and cleaning up the DB

+++ database/migrations/2018_03_21_230445_create_settings_table.php
@@ -12,15 +12,13 @@ class CreateSettingsTable extends Migration {
      */
     public function up()
     {
-        Schema::create('settings', function(Blueprint $table)
-        {
+        Schema::create('settings', function(Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade')->onUpdate('cascade');
             $table->string('name');
             $table->string('value');
             $table->timestamps();
-
         });
 
     }
